refactor(images): move `images` to `/items`
