+++ src/modules/app/app.module.ts
@@ -5,7 +5,7 @@ import { TypeOrmModule, TypeOrmModuleOptions } from '@nestjs/typeorm'
 import { environmentSchema } from '@config/environment'
 import { typeorm } from '@config/database'
 import { AuthModule } from '@modules/auth'
-import { ImagesModule } from '@modules/images'
+import { ImagesModule } from '@modules/items/images'
 import { ProfilesModule } from '@modules/profiles'
 import { UsersModule } from '@modules/users'
 import { AdaptersModule } from '@common/adapters'
