+++ src/modules/items/images/dtos/create-image.dto.ts
