+++ src/modules/items/images/dtos/index.ts
