+++ src/modules/items/images/image.entity.ts
