+++ src/modules/profiles/profiles.module.ts
@@ -4,7 +4,7 @@ import { TypeOrmModule } from '@nestjs/typeorm'
 import { Profile } from './profile.entity'
 import { ProfilesRepository } from './profiles.repository'
 
-import { ImagesModule } from '@modules/images'
+import { ImagesModule } from '@modules/items/images'
 
 @Module({
   imports: [TypeOrmModule.forFeature([Profile]), ImagesModule],
